+++ apps/web/src/features/entity/index.ts
@@ -11,6 +11,7 @@ export {
 export {
   ListEntityMetadataProvider,
   ListEntityMetadataQueryProvider,
+  ListEntityNoopMetadataProvider,
 } from './composed/list-entity/list-entity-metadata-provider';
 export { Entity } from './entity';
 export { EntityIcon as EntityRowIcon } from './extractors/entity-icon';
