+++ apps/web/src/features/property/tags/tag-sets-context.test.tsx
@@ -9,20 +9,11 @@ import { fireEvent, render, screen } from '@solidjs/testing-library';
 import type { JSX } from 'solid-js';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 import { EntityRowTags } from './EntityRowTags';
-import {
-  TagSetsProvider,
-  TagSetsQueryProvider,
-  useTagSets,
-} from './tag-sets-context';
+import { TagSetsProvider } from './tag-sets-context';
 
 const mocks = vi.hoisted(() => ({
-  useTagsQuery: vi.fn(),
   useDocTags: vi.fn(),
   useSoupDocTags: vi.fn(),
-}));
-
-vi.mock('@queries/properties/tags', () => ({
-  useTagsQuery: mocks.useTagsQuery,
 }));
 
 vi.mock('./TagPicker', () => ({
@@ -45,36 +36,12 @@ vi.mock('./useDocTags', () => ({
   useSoupDocTags: mocks.useSoupDocTags,
 }));
 
-function TagSetsConsumer() {
-  const tagSets = useTagSets();
-  return <span>{tagSets()[0]?.scope}</span>;
-}
-
 beforeEach(() => {
-  mocks.useTagsQuery.mockReset();
   mocks.useDocTags.mockReset();
   mocks.useSoupDocTags.mockReset();
-  mocks.useTagsQuery.mockReturnValue({
-    data: [{ scope: 'user', options: [] }],
-  });
 });
 
 describe('TagSetsContext', () => {
-  it('uses the provided tag sets without another query', () => {
-    const tagSets = (): TagSetResponse[] => [
-      { scope: 'team', options: [] } as TagSetResponse,
-    ];
-
-    render(() => (
-      <TagSetsProvider tagSets={tagSets}>
-        <TagSetsConsumer />
-      </TagSetsProvider>
-    ));
-
-    expect(screen.getByText('team')).toBeTruthy();
-    expect(mocks.useTagsQuery).not.toHaveBeenCalled();
-  });
-
   it('uses the real row resolver and initializes editing only on interaction', async () => {
     const tagSets = (): TagSetResponse[] => [
       {
@@ -109,7 +76,6 @@ describe('TagSetsContext', () => {
     ));
 
     expect(screen.getByText('Urgent')).toBeTruthy();
-    expect(mocks.useTagsQuery).not.toHaveBeenCalled();
     expect(mocks.useSoupDocTags).not.toHaveBeenCalled();
 
     await fireEvent.click(screen.getByRole('button'));
@@ -119,23 +85,6 @@ describe('TagSetsContext', () => {
       'document-1',
       EntityType.DOCUMENT,
       expect.any(Function)
-    );
-  });
-
-  it('makes standalone query ownership explicit', () => {
-    render(() => (
-      <TagSetsQueryProvider>
-        <TagSetsConsumer />
-      </TagSetsQueryProvider>
-    ));
-
-    expect(screen.getByText('user')).toBeTruthy();
-    expect(mocks.useTagsQuery).toHaveBeenCalledOnce();
-  });
-
-  it('fails fast outside a provider', () => {
-    expect(() => render(() => <TagSetsConsumer />)).toThrow(
-      'useTagSets can only be used under a TagSetsProvider'
     );
   });
 });
